# mobile/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "城迹 CityPulse API",
    "description": "可信、结构化、可定位的城市动态事件 API。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/events": {
      "get": {
        "tags": [
          "admin-events"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_v1_admin_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "published",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Published"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminEventDetail"
                  },
                  "title": "Response List Events Api V1 Admin Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/events/{event_id}": {
      "get": {
        "tags": [
          "admin-events"
        ],
        "summary": "Get Event",
        "operationId": "get_event_api_v1_admin_events__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin-events"
        ],
        "summary": "Update Event",
        "operationId": "update_event_api_v1_admin_events__event_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminEventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "default": "长沙",
              "title": "City"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "attribute",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Attribute"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "soonest",
                "ending_soon"
              ],
              "type": "string",
              "default": "soonest",
              "title": "Sort"
            }
          },
          {
            "name": "time_scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "upcoming",
                "past"
              ],
              "type": "string",
              "default": "upcoming",
              "title": "Time Scope"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "when",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "any",
                "today",
                "weekend"
              ],
              "type": "string",
              "default": "any",
              "title": "When"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/map": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Map Events",
        "operationId": "map_events_api_v1_events_map_get",
        "parameters": [
          {
            "name": "west",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "West"
            }
          },
          {
            "name": "south",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "South"
            }
          },
          {
            "name": "east",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "title": "East"
            }
          },
          {
            "name": "north",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "title": "North"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "default": "长沙",
              "title": "City"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminEventDetail": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "summary": {
            "type": "string",
            "maxLength": 360,
            "minLength": 1,
            "title": "Summary"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Description"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "venue_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Venue Name"
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Address"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "City"
          },
          "district": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "District"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "organizer": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Organizer"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "evidence_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "format": "uri",
            "title": "Evidence Url"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "category",
          "summary",
          "description",
          "starts_at",
          "ends_at",
          "venue_name",
          "address",
          "city",
          "district",
          "latitude",
          "longitude",
          "organizer",
          "status",
          "evidence_url",
          "id",
          "updated_at",
          "is_published"
        ],
        "title": "AdminEventDetail"
      },
...
      "AdminEventUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "summary": {
            "type": "string",
            "maxLength": 360,
            "minLength": 1,
            "title": "Summary"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Description"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "venue_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Venue Name"
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Address"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "City"
          },
          "district": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "District"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "organizer": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Organizer"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "evidence_url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "format": "uri",
            "title": "Evidence Url"
          },
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          },
          "review_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Review Note"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "category",
          "summary",
          "description",
          "starts_at",
          "ends_at",
          "venue_name",
          "address",
          "city",
          "district",
          "latitude",
          "longitude",
          "organizer",
          "status",
          "evidence_url",
          "expected_updated_at",
          "review_note",
          "is_published"
        ],
        "title": "AdminEventUpdate"
      },
...
      "EventCategory": {
        "type": "string",
        "enum": [
          "performance",
          "sports",
          "exhibition",
          "festival",
          "market",
          "public_culture",
          "pop_up",
          "seasonal"
        ],
        "title": "EventCategory"
      },
...
      "EventDetailResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EventDetail"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "EventDetailResponse"
      },
      "EventPage": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "EventPage"
      },
      "EventStatus": {
        "type": "string",
        "enum": [
          "announced",
          "on_sale",
          "sold_out",
          "postponed",
          "cancelled",
          "ended"
        ],
        "title": "EventStatus"
      },
      "EventSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "organizer": {
            "type": "string",
            "title": "Organizer"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "last_verified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Verified At"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "attributes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attributes"
          },
          "is_ad": {
            "type": "boolean",
            "title": "Is Ad"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "is_new": {
            "type": "boolean",
            "title": "Is New"
          },
          "is_ending_soon": {
            "type": "boolean",
            "title": "Is Ending Soon"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "category",
          "summary",
          "starts_at",
          "ends_at",
          "location",
          "organizer",
          "status",
          "last_verified_at",
          "confidence",
          "attributes",
          "is_ad",
          "is_demo",
          "is_new",
          "is_ending_soon"
        ],
        "title": "EventSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MapEventsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MapEvent"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "MapEventsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}